Add search through more fields in compendium.
Each derivative of CompendiumItem can use addSearchable(string) to
define what it should search on.  This allows the compendium search to
get the results easily.

Implementations to search through names, descriptions, effects, passive
effects, active effects, and so on are implemented for most (if not all)
CompendiumItem derives that use human-readable strings.